requirements.txt
torch
torchvision
opencv-python
Pillow
facenet-pytorch
numpy
requests